+++ core-bundle/src/Resources/contao/dca/tl_page.php
@@ -18,6 +18,7 @@ $GLOBALS['TL_DCA']['tl_page'] = array
 		'dataContainer'               => 'Table',
 		'ctable'                      => array('tl_article'),
 		'enableVersioning'            => true,
+		'markAsCopy'                  => 'title',
 		'onload_callback' => array
 		(
 			array('tl_page', 'checkPermission'),